requirements.txt
Django==2.1.3
djangorestframework==3.9.4
psycopg2>=2.7.5,<2.8.0
